Reworking styles and logic
