+++ modules/ToDoLogic.js
@@ -4,7 +4,6 @@ import '../src/style.css';
 import dots from '../src/images/dots-icon.png';
 import reset from '../src/images/reset-icon.png';
 
-const $input = document.querySelector('.input');
 const $listTasks = document.querySelector('.list-tasks');
 const $titleContainer = document.querySelector('.title-container');
 
@@ -12,65 +11,33 @@ const img = document.createElement('img');
 $titleContainer.appendChild(img);
 img.src = reset;
 img.classList.add('reset');
-const $reset = document.querySelector('.reset');
-
-class Task {
-  constructor(description) {
-    this.description = description;
-    this.completed = false;
-    this.index = 0;
-  }
-}
+/* const $reset = document.querySelector('.reset'); */
+const tasks = [
+  {
+    description: 'An example description in string',
+    completed: false,
+    index: 0,
+  },
+  {
+    description: 'An example description in string',
+    completed: false,
+    index: 1,
+  },
+];
 
 export class ToDo {
   constructor() {
-    this.list = [];
-    this.counter = 0;
+    this.list = tasks;
   }
 
-  handleKeyPress = (event) => {
-    const keyPressed = event.key;
-
-    if (keyPressed === 'Enter') {
-      const task = new Task($input.value);
-      if (this.list.length <= 0) {
-        this.list.push(task);
-      } else {
-        this.list.push(task);
-        this.counter += this.counter + 1;
-      }
+  loadTasks = () => {
+    tasks.forEach((task) => {
       $listTasks.innerHTML += `
-            <li class="task"><div><input type="checkbox" name="" id=""><span>${task.description}</span></div> <img class="task-img" src="${dots}" alt="options">
+            <li class="task"><div><input type="checkbox" name="" id="">
+            <span>${task.description}</span>
+            </div> <img class="task-img" src="${dots}" alt="options">
             </li>
             `;
-    }
+    });
   }
-
-    executeInputFunctioning = () => {
-      $input.addEventListener('keydown', (e) => {
-        if ($input.value !== '') {
-          this.handleKeyPress(e);
-        }
-        if (e.key === 'Enter') {
-          $input.value = '';
-        }
-      });
-    }
-
-    resetTasks = () => {
-      $reset.addEventListener('click', () => {
-        const $task = document.querySelectorAll('.task');
-        $task.forEach((task) => task.remove());
-        this.list = [];
-      });
-    }
-
-    clearAllButton = () => {
-      const $clearButton = document.getElementById('clear-button');
-      $clearButton.addEventListener('click', () => {
-        const $task = document.querySelectorAll('.task');
-        $task.forEach((task) => task.remove());
-        this.list = [];
-      });
-    }
 }
