+++ src/index.js
@@ -2,6 +2,4 @@ import { ToDo } from '../modules/ToDoLogic.js';
 
 const toDo = new ToDo();
 
-toDo.executeInputFunctioning();
-toDo.resetTasks();
-toDo.clearAllButton();
+toDo.loadTasks();
